# swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/eugenio.ciaglia/Fastify/1.0.0
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: events
    description: Event domain
paths:
  /eventi:
    get:
      tags:
        - events
      summary: Search all events.
      operationId: readEvents
      description: |
        Can search for all events available.
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventItem'
        '400':
          description: bad input parameter
    post:
      tags:
        - events
      summary: adds an event
      operationId: createEvent
      description: Adds an event to the system
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: an existing item already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventItem'
        description: Inventory item to add
  /eventi/{id}:
    put:
      tags:
        - events
      summary: update an event informations given event id.
      operationId: updateEvent
      description: Update an event to the system
      parameters:
        - name: id
          in: path
          description: Event ID
          required: true
          schema:
            type: string
            format: uuid
            example: d290f1ee-6c54-4b01-90e6-d701748f0851
      responses:
        '201':
          description: item updated
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: an existing item already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventItem'
        description: Inventory item to update
    delete:
      tags:
        - events
      summary: Delete an event given event id.
      operationId: deleteEvent
      description: Delete an event to the system
      parameters:
        - name: id
          in: path
          description: Event ID
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "61ae4060f3851abf6cee6067"
      responses:
        '201':
          description: item deleted
        '400':
          description: "invalid id, object doesn't exists"
        '409':
          description: an existing item already exists
components:
  schemas:
    EventItem:
      type: object
      required:
        - id
        - title
        - description
        - startDate
        - endDate
      properties:
        id:
          type: string
          minLength: 24
          maxLength: 24
          example: "61ae4060f3851abf6cee6067"
        title:
          type: string
          example: Concerto
        description:
          type: string
          example: Concerto Hans Zimmer - Bologna
        startDate:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        endDate:
          type: string
          format: date-time
          example: '2016-08-29T10:12:33.001Z'